requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asyncio==3.4.3
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
customtkinter==4.6.3
cycler==0.12.1
darkdetect==0.7.1
distro==1.8.0
exceptiongroup==1.2.0
fastapi==0.105.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.2
gradio==4.31.2
gradio_client==0.16.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
load-dotenv==0.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.26.2
openai==1.30.1
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pdfminer.six==20231228
Pillow==9.2.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyodbc==5.0.1
pyparsing==3.1.1
PyPDF2==3.0.1
pystray==0.19.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
ruff==0.4.4
schedule==1.1.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
tiktoken==0.5.2
tk==0.1.0
tomlkit==0.12.0
toolz==0.12.0
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.24.0.post1
websockets==11.0.3
